+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Argamak now requires Gleam v0.33 or later.
 - The `tensor` module's `concat` function now returns an `AxisNotFound` error
   when the given find function is `False` for every `Axis`.